replace StaticSolver files
