+++ Sofa/Component/ODESolver/Backward/CMakeLists.txt
@@ -12,7 +12,6 @@ set(HEADER_FILES
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewmarkImplicitSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonRaphsonSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonStatus.h
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticOdeSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/VariationalSymplecticSolver.h
     
@@ -31,7 +30,6 @@ set(SOURCE_FILES
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/EulerImplicitSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewmarkImplicitSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonRaphsonSolver.cpp
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticOdeSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/StaticSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/VariationalSymplecticSolver.cpp
 )
