--- examples/Component/ODESolver/Backward/StaticOdeSolver.scn
@@ -1,41 +1,0 @@
-<Node name="root" gravity="-1.8 0 100"  dt="1">
-    <RequiredPlugin name="Sofa.Component.Constraint.Projective"/> <!-- Needed to use components [FixedProjectiveConstraint] -->
-    <RequiredPlugin name="Sofa.Component.IO.Mesh"/> <!-- Needed to use components [MeshGmshLoader MeshOBJLoader] -->
-    <RequiredPlugin name="Sofa.Component.LinearSolver.Iterative"/> <!-- Needed to use components [CGLinearSolver] -->
-    <RequiredPlugin name="Sofa.Component.Mapping.Linear"/> <!-- Needed to use components [BarycentricMapping] -->
-    <RequiredPlugin name="Sofa.Component.Mass"/> <!-- Needed to use components [UniformMass] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Backward"/> <!-- Needed to use components [StaticSolver] -->
-    <RequiredPlugin name="Sofa.Component.SolidMechanics.FEM.Elastic"/> <!-- Needed to use components [TetrahedronFEMForceField] -->
-    <RequiredPlugin name="Sofa.Component.SolidMechanics.Spring"/> <!-- Needed to use components [MeshSpringForceField] -->
-    <RequiredPlugin name="Sofa.Component.StateContainer"/> <!-- Needed to use components [MechanicalObject] -->
-    <RequiredPlugin name="Sofa.Component.Topology.Container.Constant"/> <!-- Needed to use components [MeshTopology] -->
-    <RequiredPlugin name="Sofa.Component.Visual"/> <!-- Needed to use components [VisualStyle] -->
-    <RequiredPlugin name="Sofa.GL.Component.Rendering3D"/> <!-- Needed to use components [OglModel] -->
-    <DefaultAnimationLoop/>
-
-    <VisualStyle displayFlags="hideBehaviorModels hideCollisionModels hideMappings hideForceFields" />
-    <Node name="Reference">
-        <MeshOBJLoader name="meshLoader_1" filename="mesh/truthcylinder1-bent.obj" scale="0.95" handleSeams="1" />
-        <OglModel src="@meshLoader_1" dx="0" dy="-1" dz="0" color="green" />
-    </Node>
-
-    <Node name="CoFEM">
-        <NewtonRaphsonSolver name="newtonSolver" printLog="true" maxNbIterationsNewton="100" maxNbIterationsLineSearch="1" warnWhenLineSearchFails="false"/>
-        <StaticOdeSolver newtonSolver="@../newtonSolver" printLog="true"/>
-        <EigenSimplicialLDLT template="CompressedRowSparseMatrixMat3x3"/>
-        <MeshGmshLoader name="loader" filename="mesh/truthcylinder1.msh"/>
-        <MeshTopology src="@loader" />
-        <MechanicalObject src="@loader" dx="15" />
-        <MeshMatrixMass totalMass="15" />
-        <FixedProjectiveConstraint indices="0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 268 269 270 271 343 345" />
-        <TetrahedronFEMForceField name="FEM" youngModulus="1116" poissonRatio="0.49" method="large" />
-        <Node>
-            <MeshOBJLoader name="meshLoader_3" filename="mesh/truthcylinder1.obj" handleSeams="1" />
-            <OglModel name="Visual" src="@meshLoader_3" color="cyan" dx="15" />
-            <BarycentricMapping input="@.." output="@Visual" />
-        </Node>
-    </Node>
-
-
-
-</Node>
