+++ examples/RegressionStateScenes.regression-tests
@@ -9,7 +9,7 @@
 # USE UNDERSCORES ("_") INSTEAD.
 #
 # The file contains the path to the reference folder. This path should be relative to this folder path. The folder
-# should be structured as the folder containig this file and each reference of each scene should have a relative path to
+# should be structured as the folder containing this file and each reference of each scene should have a relative path to
 # the reference folder equivalent to the relative path of the corresponding scene in this folder. e.g. for the scene in
 # path/to/MyScene.scn, a reference with the same name as the scene should be found in referenceFolder/path/to/
 # --> INFO : a special string can be use while specifying this path: `$REGRESSION_DIR`. If used at the beginning of the
